+++ src/mss.js
@@ -1,4 +1,3 @@
-const uuid = require('uuid');
 const cookieSession = require('cookie-session');
 const cookieParser = require('cookie-parser');
 const express = require('express');
@@ -108,31 +107,6 @@ const creeServeur = (
   });
 
   app.get(
-    '/initialisationMotDePasse/:idReset',
-    middleware.aseptise('idReset'),
-    async (requete, reponse) => {
-      const { idReset } = requete.params;
-
-      const pasUnUUID = !uuid.validate(idReset);
-      if (pasUnUUID) {
-        reponse.status(400).send(`UUID requis`);
-        return;
-      }
-
-      const utilisateur = await depotDonnees.utilisateurAFinaliser(idReset);
-      if (!utilisateur) {
-        reponse
-          .status(404)
-          .send(`Identifiant d'initialisation de mot de passe inconnu`);
-        return;
-      }
-
-      requete.session.token = utilisateur.genereToken();
-      reponse.render('motDePasse/edition', { utilisateur });
-    }
-  );
-
-  app.get(
     '/espacePersonnel',
     middleware.verificationAcceptationCGU,
     (_requete, reponse) => {
@@ -157,7 +131,7 @@ const creeServeur = (
     }
   );
 
-  app.use('', routesNonConnectePage({ middleware, referentiel }));
+  app.use('', routesNonConnectePage({ depotDonnees, middleware, referentiel }));
 
   app.use(
     '/api',
