+++ src/routes/nonConnecte/routesNonConnectePage.js
@@ -1,10 +1,11 @@
 const express = require('express');
+const uuid = require('uuid');
 const {
   estUrlLegalePourRedirection,
   construisUrlAbsolueVersPage,
 } = require('../../http/redirection');
 
-const routesNonConnectePage = ({ middleware, referentiel }) => {
+const routesNonConnectePage = ({ depotDonnees, middleware, referentiel }) => {
   const routes = express.Router();
 
   routes.get('/', (_requete, reponse) => {
@@ -72,6 +73,31 @@ const routesNonConnectePage = ({ middleware, referentiel }) => {
     middleware.suppressionCookie,
     (_requete, reponse) => {
       reponse.render('reinitialisationMotDePasse');
+    }
+  );
+
+  routes.get(
+    '/initialisationMotDePasse/:idReset',
+    middleware.aseptise('idReset'),
+    async (requete, reponse) => {
+      const { idReset } = requete.params;
+
+      const pasUnUUID = !uuid.validate(idReset);
+      if (pasUnUUID) {
+        reponse.status(400).send(`UUID requis`);
+        return;
+      }
+
+      const utilisateur = await depotDonnees.utilisateurAFinaliser(idReset);
+      if (!utilisateur) {
+        reponse
+          .status(404)
+          .send(`Identifiant d'initialisation de mot de passe inconnu`);
+        return;
+      }
+
+      requete.session.token = utilisateur.genereToken();
+      reponse.render('motDePasse/edition', { utilisateur });
     }
   );
